update new ima

+++ rabbitmq/ec2-consumer.js
@@ -1,3 +1,4 @@
+import os from 'os';
 
 import {
     createEC2AndRunConsumer,
@@ -17,13 +18,18 @@ import {
     sendMessage
 } from './mq.js';
 
+const isSkip = (os.platform() === 'darwin'); // which is mac for local development
+
 async function main() {
     let channel = await getChannel(QUEUE_EC2);
     consume(channel, QUEUE_EC2, 1, callback);
     async function callback(msg) {
         console.log('come in');
         try {
-            const instanceIDs = await createEC2AndRunConsumer(EC2_COUNT);
+            let instanceIDs;
+            if (!isSkip) {
+                instanceIDs = await createEC2AndRunConsumer(EC2_COUNT);
+            }
             await checkQueueISEmpty(instanceIDs);
             channel.ack(msg);
         } catch (error) {
@@ -33,6 +39,11 @@ async function main() {
     }
 }
 
+async function sendNotification() {
+    const diffChannel = await getChannel(QUEUE_DIFF);
+    sendMessage(diffChannel, QUEUE_DIFF)
+}
+
 async function checkQueueISEmpty(instanceIDs) {
     const interval = setInterval(async () => {
         let channel = await getChannel();
@@ -40,9 +51,10 @@ async function checkQueueISEmpty(instanceIDs) {
         console.log('unhandle messages:', messageCount);
         if (messageCount === 0) {
             clearInterval(interval);
-            await deleteEC2(instanceIDs);
-            const diffChannel = await getChannel(QUEUE_DIFF);
-            sendMessage(diffChannel, QUEUE_DIFF)
+            if (!isSkip) {
+                await deleteEC2(instanceIDs);
+            }
+            await sendNotification();
         }
     }, 1000 * 10);
 }
